refactored

+++ src/main/java/com/qa/calculator/Add.java
@@ -2,7 +2,7 @@ package com.qa.calculator;
 
 public class Add {
 
-	public int add(int num1, int num2) {
+	public double add(double num1, double num2) {
 		return num1 + num2;
 	}
