+++ src/main/java/com/qa/calculator/Divide.java
@@ -2,7 +2,7 @@ package com.qa.calculator;
 
 public class Divide {
 
-	public int divide(int number1, int number2) {
+	public double divide(double number1, double number2) {
 		return number1/number2;
 	}
