+++ src/main/java/com/qa/calculator/Multiply.java
@@ -2,7 +2,7 @@ package com.qa.calculator;
 
 public class Multiply {
 
-	public int multiply(int number1, int number2) {
+	public double multiply(double number1, double number2) {
 		return number1*number2;
 	}
