+++ src/main/java/com/qa/calculator/Subtract.java
@@ -2,7 +2,7 @@ package com.qa.calculator;
 
 public class Subtract {
 
-	public int subtract(int number1, int number2) {
+	public double subtract(double number1, double number2) {
 		return number1 - number2;
 	}
